upload , save, delete images on S3 bucket

+++ src/restaurants/restaurants.controller.ts
@@ -6,13 +6,16 @@ import {
   Param,
   Post,
   Put,
-  Query
+  Query,
+  UploadedFiles,
+  UseInterceptors,
 } from '@nestjs/common';
 import { Query as ExpressQuery } from 'express-serve-static-core';
 import { RestaurantsService } from './restaurants.service';
 import { Restaurant } from './schemas/restaurants.schema';
 import { CreateRestaurantDto } from './dto/create-restaurant.dto';
 import { UpdateRestaurantDto } from './dto/update-restaurant.dto';
+import { FilesInterceptor } from '@nestjs/platform-express';
 
 @Controller('restaurants')
 export class RestaurantsController {
@@ -48,12 +51,38 @@ export class RestaurantsController {
   async deleteRestaurant(
     @Param('id') id: string,
   ): Promise<{ deleted: Boolean }> {
-    await this.restaurantService.findById(id);
-    const res = this.restaurantService.deleteById(id);
-    if (res) {
+    const restaurant = await this.restaurantService.findById(id);
+
+    const isDeleted = await this.restaurantService.deleteImages(
+      restaurant.images,
+    );
+
+    if (isDeleted) {
+      const res = this.restaurantService.deleteById(id);
+      if (res) {
+        return {
+          deleted: true,
+        };
+      }
+    } else {
       return {
-        deleted: true,
+        deleted: false,
       };
     }
+  }
+
+  @Put('upload/:id')
+  @UseInterceptors(FilesInterceptor('files'))
+  async uploadFile(
+    @Param('id') id: string,
+    @UploadedFiles() files: Array<Express.Multer.File>,
+  ) {
+    console.log(files);
+    console.log(id);
+
+    await this.restaurantService.findById(id);
+
+    const res = await this.restaurantService.uploadImages(id, files);
+    return res;
   }
 }
